Remove _Continuation protocol and duplicate the relevant code.
