+++ stdlib/public/Concurrency/CMakeLists.txt
@@ -37,7 +37,6 @@ add_swift_target_library(swift_Concurrency ${SWIFT_STDLIB_LIBRARY_BUILD_TYPES} I
   Actor.cpp
   Actor.swift
   CheckedContinuation.swift
-  ContinuationProtocol.swift
   GlobalExecutor.cpp
   AsyncIteratorProtocol.swift
   AsyncSequence.swift
